+++ tests/hook_tests.rs
@@ -17,7 +17,7 @@ fn hook_for_simple_callee<'p, B: Backend>(
 ) -> Result<ReturnValue<B::BV>> {
     assert_eq!(call.get_arguments().len(), 2);
     Ok(ReturnValue::Return(
-        state.bv_from_u32(5, state.size(&call.get_type()) as u32),
+        state.bv_from_u32(5, state.size(&call.get_type())),
     ))
 }
 
@@ -65,7 +65,7 @@ fn target_hook<'p, B: Backend>(
 ) -> Result<ReturnValue<B::BV>> {
     assert_eq!(call.get_arguments().len(), 2);
     Ok(ReturnValue::Return(
-        state.bv_from_u32(5, state.size(&call.get_type()) as u32),
+        state.bv_from_u32(5, state.size(&call.get_type())),
     ))
 }
